[CPDNPQ-2810] Extract mermaid JS bundle
The Mermaid JS library we use isn't compatible with some older browsers that we're mandated to support, e.g. Safari 16.1 and lower

Mermaid is only used on API guidance pages, which have drastically less traffic and it's reasonable to assume are mostly access by developers i.e. legacy browsers less likely

Only loading mermaid on those pages is a quicker route to fixing the main JavaScript bundle in older browsers than trying to get Mermaid to run in them
